Memgpt

+++ minichain/tools/deploy_static.py
@@ -15,9 +15,17 @@ async def deploy_static_website(
     new_public_path = path.split("/")[-1]
     target, v = os.path.join(settings.SERVE_PATH, new_public_path ), 0
     while os.path.exists(target):
-        new_public_path = path.split("/")[-1] + f"_{v}"
-        target = os.path.join(settings.SERVE_PATH, new_public_path )
-        v += 1
-    shutil.copytree(path, target)
+        if os.path.isfile(path):
+            # delete the file if it exists
+            os.remove(target)
+        else:
+            new_public_path = path.split("/")[-1] + f"_{v}"
+            target = os.path.join(settings.SERVE_PATH, new_public_path )
+            v += 1
+    if os.path.isfile(path):
+        os.makedirs(os.path.dirname(target), exist_ok=True)
+        shutil.copyfile(path, target)
+    else:
+        shutil.copytree(path, target)
     public_url = settings.SERVE_URL + new_public_path
     return f"Your file(s) are now available [here]({public_url})"
